Added tests for fixed problem #366 with passing context from adapter client/server. Removed package capturecontext
