+++ pkg/registry/core/adapters/ns_registry.go
@@ -19,8 +19,9 @@ package adapters
 import (
 	"context"
 	"errors"
-	"github.com/networkservicemesh/sdk/pkg/registry/core/streamcontext"
 	"io"
+
+	"github.com/networkservicemesh/sdk/pkg/registry/core/streamcontext"
 
 	"github.com/networkservicemesh/sdk/pkg/registry/core/next"
 
@@ -37,19 +38,20 @@ type networkServiceRegistryServer struct {
 }
 
 func (n *networkServiceRegistryServer) Register(ctx context.Context, request *registry.NetworkService) (*registry.NetworkService, error) {
-	doneCtx := withDoneContext(ctx)
+	doneCtx := withCapturedContext(ctx)
 	ns, err := n.client.Register(doneCtx, request)
 	if err != nil {
 		return nil, err
 	}
-	if !isDoneContext(doneCtx) {
+	lastCtx := getCapturedContext(doneCtx)
+	if lastCtx == nil {
 		return ns, nil
 	}
-	return next.NetworkServiceRegistryServer(ctx).Register(lastContext(doneCtx), request)
+	return next.NetworkServiceRegistryServer(ctx).Register(lastCtx, request)
 }
 
 func (n *networkServiceRegistryServer) Find(query *registry.NetworkServiceQuery, s registry.NetworkServiceRegistry_FindServer) error {
-	doneCtx := withDoneContext(s.Context())
+	doneCtx := withCapturedContext(s.Context())
 	client, err := n.client.Find(doneCtx, query)
 	if err != nil {
 		return err
@@ -73,27 +75,29 @@ func (n *networkServiceRegistryServer) Find(query *registry.NetworkServiceQuery,
 			return err
 		}
 	}
-	if isDoneContext(doneCtx) {
-		return next.NetworkServiceRegistryServer(s.Context()).Find(query, streamcontext.NetworkServiceRegistryFindServer(lastContext(doneCtx), s))
+	lastCtx := getCapturedContext(doneCtx)
+	if lastCtx != nil {
+		return next.NetworkServiceRegistryServer(s.Context()).Find(query, streamcontext.NetworkServiceRegistryFindServer(lastCtx, s))
 	}
 	return nil
 }
 
 func (n *networkServiceRegistryServer) Unregister(ctx context.Context, request *registry.NetworkService) (*empty.Empty, error) {
-	doneCtx := withDoneContext(ctx)
+	doneCtx := withCapturedContext(ctx)
 	ns, err := n.client.Unregister(doneCtx, request)
 	if err != nil {
 		return nil, err
 	}
-	if !isDoneContext(doneCtx) {
+	lastCtx := getCapturedContext(doneCtx)
+	if lastCtx == nil {
 		return ns, nil
 	}
-	return next.NetworkServiceRegistryServer(ctx).Unregister(lastContext(doneCtx), request)
+	return next.NetworkServiceRegistryServer(ctx).Unregister(lastCtx, request)
 }
 
 // NetworkServiceClientToServer - returns a registry.NetworkServiceRegistryClient wrapped around the supplied client
 func NetworkServiceClientToServer(client registry.NetworkServiceRegistryClient) registry.NetworkServiceRegistryServer {
-	return &networkServiceRegistryServer{client: next.NewNetworkServiceRegistryClient(client, &nsDoneClient{})}
+	return &networkServiceRegistryServer{client: next.NewNetworkServiceRegistryClient(client, NewNetworkServiceRegistryClient())}
 }
 
 var _ registry.NetworkServiceRegistryServer = &networkServiceRegistryServer{}
@@ -103,20 +107,21 @@ type networkServiceRegistryClient struct {
 }
 
 func (n *networkServiceRegistryClient) Register(ctx context.Context, in *registry.NetworkService, _ ...grpc.CallOption) (*registry.NetworkService, error) {
-	doneCtx := withDoneContext(ctx)
+	doneCtx := withCapturedContext(ctx)
 	ns, err := n.server.Register(doneCtx, in)
 	if err != nil {
 		return nil, err
 	}
-	if !isDoneContext(doneCtx) {
+	lastCtx := getCapturedContext(doneCtx)
+	if lastCtx == nil {
 		return ns, nil
 	}
-	return next.NetworkServiceRegistryClient(ctx).Register(lastContext(doneCtx), in)
+	return next.NetworkServiceRegistryClient(ctx).Register(lastCtx, in)
 }
 
-func (n *networkServiceRegistryClient) Find(ctx context.Context, in *registry.NetworkServiceQuery, opts ...grpc.CallOption) (registry.NetworkServiceRegistry_FindClient, error) {
+func (n *networkServiceRegistryClient) Find(ctx context.Context, in *registry.NetworkServiceQuery, _ ...grpc.CallOption) (registry.NetworkServiceRegistry_FindClient, error) {
 	ch := make(chan *registry.NetworkService, channelSize)
-	doneCtx := withDoneContext(ctx)
+	doneCtx := withCapturedContext(ctx)
 	s := streamchannel.NewNetworkServiceFindServer(doneCtx, ch)
 	if in != nil && in.Watch {
 		go func() {
@@ -129,8 +134,9 @@ func (n *networkServiceRegistryClient) Find(ctx context.Context, in *registry.Ne
 			return nil, err
 		}
 	}
-	if isDoneContext(doneCtx) {
-		_, err := next.NetworkServiceRegistryClient(ctx).Find(lastContext(doneCtx), in)
+	lastCtx := getCapturedContext(doneCtx)
+	if lastCtx != nil {
+		_, err := next.NetworkServiceRegistryClient(ctx).Find(lastCtx, in)
 		if err != nil {
 			return nil, err
 		}
@@ -139,20 +145,21 @@ func (n *networkServiceRegistryClient) Find(ctx context.Context, in *registry.Ne
 }
 
 func (n *networkServiceRegistryClient) Unregister(ctx context.Context, in *registry.NetworkService, _ ...grpc.CallOption) (*empty.Empty, error) {
-	doneCtx := withDoneContext(ctx)
+	doneCtx := withCapturedContext(ctx)
 	ns, err := n.server.Unregister(doneCtx, in)
 	if err != nil {
 		return nil, err
 	}
-	if !isDoneContext(doneCtx) {
+	lastCtx := getCapturedContext(doneCtx)
+	if lastCtx == nil {
 		return ns, nil
 	}
-	return next.NetworkServiceRegistryClient(ctx).Unregister(lastContext(doneCtx), in)
+	return next.NetworkServiceRegistryClient(ctx).Unregister(lastCtx, in)
 }
 
 var _ registry.NetworkServiceRegistryClient = &networkServiceRegistryClient{}
 
 // NetworkServiceServerToClient - returns a registry.NetworkServiceRegistryServer wrapped around the supplied server
 func NetworkServiceServerToClient(server registry.NetworkServiceRegistryServer) registry.NetworkServiceRegistryClient {
-	return &networkServiceRegistryClient{server: next.NewNetworkServiceRegistryServer(server, &nsDoneServer{})}
+	return &networkServiceRegistryClient{server: next.NewNetworkServiceRegistryServer(server, NewNetworkServiceRegistryServer())}
 }
