+++ pkg/registry/core/adapters/ns_registry_test.go
@@ -20,6 +20,10 @@ import (
 	"context"
 	"testing"
 
+	"github.com/networkservicemesh/sdk/pkg/registry/core/next"
+	"github.com/networkservicemesh/sdk/pkg/registry/core/streamcontext"
+	"github.com/networkservicemesh/sdk/pkg/registry/utils/checks/checkcontext"
+
 	"github.com/golang/protobuf/ptypes/empty"
 	"github.com/networkservicemesh/api/pkg/api/registry"
 	"github.com/stretchr/testify/require"
@@ -31,7 +35,7 @@ import (
 
 type echoNetworkServiceClient struct{}
 
-func (t *echoNetworkServiceClient) Register(ctx context.Context, in *registry.NetworkService, opts ...grpc.CallOption) (*registry.NetworkService, error) {
+func (t *echoNetworkServiceClient) Register(_ context.Context, in *registry.NetworkService, _ ...grpc.CallOption) (*registry.NetworkService, error) {
 	return in, nil
 }
 
@@ -95,4 +99,78 @@ func adaptNetworkServiceClientToServerFewTimes(n int, client registry.NetworkSer
 	}
 
 	return s
+}
+
+type contextKeyType string
+
+const testKey contextKeyType = "TestContext"
+
+type writeNSServer struct {
+}
+
+func (w *writeNSServer) Register(ctx context.Context, service *registry.NetworkService) (*registry.NetworkService, error) {
+	ctx = context.WithValue(ctx, testKey, true)
+	return next.NetworkServiceRegistryServer(ctx).Register(ctx, service)
+}
+
+func (w *writeNSServer) Find(query *registry.NetworkServiceQuery, server registry.NetworkServiceRegistry_FindServer) error {
+	ctx := context.WithValue(server.Context(), testKey, true)
+	return next.NetworkServiceRegistryServer(server.Context()).Find(query, streamcontext.NetworkServiceRegistryFindServer(ctx, server))
+}
+
+func (w *writeNSServer) Unregister(ctx context.Context, service *registry.NetworkService) (*empty.Empty, error) {
+	ctx = context.WithValue(ctx, testKey, true)
+	return next.NetworkServiceRegistryServer(ctx).Unregister(ctx, service)
+}
+
+func TestNSClientPassingContext(t *testing.T) {
+	n := next.NewNetworkServiceRegistryClient(adapters.NetworkServiceServerToClient(&writeNSServer{}), checkcontext.NewNSClient(t, func(t *testing.T, ctx context.Context) {
+		if _, ok := ctx.Value(testKey).(bool); !ok {
+			t.Error("Context ignored")
+		}
+	}))
+
+	_, err := n.Register(context.Background(), nil)
+	require.NoError(t, err)
+
+	_, err = n.Find(context.Background(), nil)
+	require.NoError(t, err)
+
+	_, err = n.Unregister(context.Background(), nil)
+	require.NoError(t, err)
+}
+
+type writeNSClient struct {
+}
+
+func (w *writeNSClient) Register(ctx context.Context, in *registry.NetworkService, opts ...grpc.CallOption) (*registry.NetworkService, error) {
+	ctx = context.WithValue(ctx, testKey, true)
+	return next.NetworkServiceRegistryClient(ctx).Register(ctx, in, opts...)
+}
+
+func (w *writeNSClient) Find(ctx context.Context, in *registry.NetworkServiceQuery, opts ...grpc.CallOption) (registry.NetworkServiceRegistry_FindClient, error) {
+	ctx = context.WithValue(ctx, testKey, true)
+	return next.NetworkServiceRegistryClient(ctx).Find(ctx, in, opts...)
+}
+
+func (w *writeNSClient) Unregister(ctx context.Context, in *registry.NetworkService, opts ...grpc.CallOption) (*empty.Empty, error) {
+	ctx = context.WithValue(ctx, testKey, true)
+	return next.NetworkServiceRegistryClient(ctx).Unregister(ctx, in, opts...)
+}
+
+func TestNSServerPassingContext(t *testing.T) {
+	n := next.NewNetworkServiceRegistryServer(adapters.NetworkServiceClientToServer(&writeNSClient{}), checkcontext.NewNSServer(t, func(t *testing.T, ctx context.Context) {
+		if _, ok := ctx.Value(testKey).(bool); !ok {
+			t.Error("Context ignored")
+		}
+	}))
+
+	_, err := n.Register(context.Background(), nil)
+	require.NoError(t, err)
+
+	err = n.Find(nil, streamcontext.NetworkServiceRegistryFindServer(context.Background(), nil))
+	require.NoError(t, err)
+
+	_, err = n.Unregister(context.Background(), nil)
+	require.NoError(t, err)
 }
