+++ pkg/registry/core/adapters/nse_capturecontext_server.go
@@ -14,7 +14,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package capturecontext
+package adapters
 
 import (
 	"context"
@@ -42,9 +42,9 @@ func (c *contextNSEServer) Unregister(ctx context.Context, in *registry.NetworkS
 	return next.NetworkServiceEndpointRegistryServer(ctx).Unregister(ctx, in)
 }
 
-// NewNetworkServiceEndpointRegistryServer - creates a new registry.NetworkServiceEndpointRegistryServer chain element that can store
-// current context for further use with CapturedContext function.
-// For this purpose it's need to use WithCapturedContext in one of the previous chain element.
+// NewNetworkServiceEndpointRegistryServer - creates a new registry.NetworkServiceEndpointRegistryServer chain element
+// that adds into context reference to current context for passing it from the adapter server/client to the next
+// client/server to avoid the problem with losing values from adapted server/client context.
 func NewNetworkServiceEndpointRegistryServer() registry.NetworkServiceEndpointRegistryServer {
 	return &contextNSEServer{}
 }
